// apps/orchestrator/src/types/database-queue-types.ts
export interface JobOption {
    attempts?: number;
    delay?: number;
    removeOnComplete?: number;
    removeOnFail?: number;
}

export enum QueueJobTypes {
    UPDATE_GAME_SESSION = 'UPDATE_GAME_SESSION',
    UPDATE_QUIZ = 'UPDATE_QUIZ',
    UPDATE_PARTICIPANT = 'UPDATE_PARTICIPANT',
    UPDATE_SPECTATOR = 'UPDATE_SPECTATOR',
    CREATE_CHAT_MESSAGE = 'CREATE_CHAT_MESSAGE',
    CREATE_CHAT_REACTION = 'CREATE_CHAT_REACTION',
    CREATE_PARTICIPANT_RESPONSE = 'CREATE_PARTICIPANT_RESPONSE',
    CREATE_LIFELINE_USAGE = 'CREATE_LIFELINE_USAGE',
    UPDATE_QUESTION = 'UPDATE_QUESTION',
    LIFELINE_EXPIRY = 'LIFELINE_EXPIRY',
}